vise

Struct Registry

pub struct Registry { /* private fields */ }

Metrics registry.

A registry collects Metrics and Collectors defined in an app and libs the app depends on. Then, these metrics can be scraped by calling Self::encode().

Collecting metrics

You can include Metrics and Collectors to a registry manually using Self::register_metrics() and Self::register_collector(). However, this can become untenable for large apps with a complex dependency graph. As an alternative, you may use register attributes to mark Metrics and Collectors that should be present in the registry, and then initialize the registry with MetricsCollection methods.

use vise::{Buckets, Global, Histogram, Metrics, MetricsCollection, Registry, Unit};
use std::time::Duration;

#[derive(Debug, Metrics)]
#[metrics(prefix = "my_app")]
pub(crate) struct AppMetrics {
    /// Latency of requests served by the app.
    #[metrics(buckets = Buckets::LATENCIES, unit = Unit::Seconds)]
    pub request_latency: Histogram<Duration>,
}

#[vise::register]
// ^ Registers this instance for use with `MetricsCollection::collect()`
pub(crate) static APP_METRICS: Global<AppMetrics> = Global::new();

let registry: Registry = MetricsCollection::default().collect();
// Check that the registered metric is present
let descriptor = registry
    .descriptors()
    .metric("my_app_request_latency_seconds")
    .unwrap();
assert_eq!(descriptor.metric.help, "Latency of requests served by the app");
assert_matches!(descriptor.metric.unit, Some(Unit::Seconds));

Registered metrics can be filtered. This is useful if you want to avoid exporting certain metrics under certain conditions.

let filtered_registry: Registry = MetricsCollection::default()
    .filter(|group| group.name == "AppMetrics")
    .collect();
// Do something with `filtered_registry`...

collect() will panic if a metric is redefined:

#[derive(Debug, Metrics)]
pub(crate) struct AppMetrics {
    #[metrics(unit = Unit::Bytes)]
    cache_memory_use: Gauge<u64>,
}

#[vise::register]
pub(crate) static APP_METRICS: Global<AppMetrics> = Global::new();

// Collector that provides the same (already registered!) metrics. This is
// logically incorrect; don't do this.
#[vise::register]
pub(crate) static APP_COLLECTOR: Collector<AppMetrics> = Collector::new();

let registry = MetricsCollection::default().collect(); // will panic

Implementations

impl Registry

pub fn empty() -> Self

Creates an empty registry.

pub fn descriptors(&self) -> &RegisteredDescriptors

Returns descriptors for all registered metrics.

pub fn register_metrics<M: Metrics>(&mut self, metrics: &M)

Registers a group of metrics.

pub fn register_collector<M: Metrics>( &mut self, collector: &'static Collector<M>, )

Registers a Collector.

pub fn encode<W: Write>(&self, writer: &mut W, format: Format) -> Result

Encodes all metrics in this registry to the specified text format.

Errors

Proxies formatting errors of the provided writer.