Trait vise::traits::HistogramValue

pub trait HistogramValue:
    'static
    + Copy
    + Debug {
    // Required method
    fn encode(self) -> f64;
}

Value of a Histogram.

This trait is implemented for signed and unsigned integers (i64, u64, isize, usize), f64 and Duration. To use smaller ints and floats as Histogram values, they can be converted to their larger-sized variants (e.g., i16 to i64, u32 to u64, and f32 to f64).

Required Methods

fn encode(self) -> f64

Encodes this value for exporting.

Object Safety

This trait is not object safe.