Trait vise::traits::MapLabels

pub trait MapLabels<S>: Copy {
    type Output<'a>: EncodeLabelSet
       where Self: 'a,
             S: 'a;

    // Required method
    fn map_labels<'a>(&'a self, labels: &'a S) -> Self::Output<'a>;
}

Maps a set of labels from the storage format (i.e., how labels are stored in a Family) to the encoding format, which is used when exporting metrics.

Required Associated Types

type Output<'a>: EncodeLabelSet where Self: 'a, S: 'a

Result of the mapping.

Required Methods

fn map_labels<'a>(&'a self, labels: &'a S) -> Self::Output<'a>

Performs mapping.