vise

Attribute Macro register

#[register]

Registers a Global metrics instance or Collector, so that it will be included into registries instantiated using MetricsCollection.

This macro must be placed on a static item of a type implementing CollectToRegistry.

Examples

Usage with global metrics

use vise::{Gauge, Global, Metrics};

#[derive(Debug, Metrics)]
#[metrics(prefix = "test")]
pub(crate) struct TestMetrics {
    gauge: Gauge,
}

#[vise::register]
static TEST_METRICS: Global<TestMetrics> = Global::new();

Usage with collectors

use vise::{Collector, Gauge, Global, Metrics};

#[derive(Debug, Metrics)]
#[metrics(prefix = "dynamic")]
pub(crate) struct DynamicMetrics {
    gauge: Gauge,
}

#[vise::register]
static TEST_COLLECTOR: Collector<DynamicMetrics> = Collector::new();