vise

Derive Macro EncodeLabelSet

#[derive(EncodeLabelSet)]
{
    // Attributes available to this derive:
    #[metrics]
}

Derives the EncodeLabelSet trait for a type, which encodes a set of metric labels.

The type for which the trait is derived must not have generics (lifetimes, type params or const params). The macro can be configured using #[metrics()] attributes.

Container attributes

label

Type: string

If specified, the type will be treated as a single label with the given name. This covers the common case in which a label set consists of a single label. In this case, the type also needs to implement EncodeLabelValue.

If this attribute is not specified (which is the default), a container must be a struct with named fields. A label with the matching name will be created for each field.

Field attributes

skip

Type: path to a function with fn(&FieldType) -> bool signature

This attribute works similarly to skip_serializing_if in serde – if the function it points to returns true for the field value, the field will not be encoded as a label.

Option fields are skipped by default if they are None (i.e., they use skip = Option::is_none).

unit

Type: expression evaluating to Unit

Specifies unit of measurement for a label. The unit will be added to the label name as a suffix (e.g., timeout_seconds if placed on a field named timeout). This is mostly useful for Info metrics.

Examples

Set with a single label

use derive_more::Display;
use vise::{EncodeLabelSet, EncodeLabelValue};

#[derive(Debug, Display, Clone, PartialEq, Eq, Hash)]
#[derive(EncodeLabelValue, EncodeLabelSet)]
#[metrics(label = "method")]
struct Method(&'static str);

Set with multiple labels

#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EncodeLabelSet)]
struct Labels {
    /// Label that is skipped when empty.
    #[metrics(skip = str::is_empty)]
    name: &'static str,
    /// Numeric label.
    num: u8,
}

Set for info metric

use vise::{EncodeLabelSet, DurationAsSecs, Unit};
use std::time::Duration;

#[derive(Debug, EncodeLabelSet)]
struct InfoLabels {
    /// Simple label.
    version: &'static str,
    /// Label with a unit.
    #[metrics(unit = Unit::Seconds)]
    request_timeout: DurationAsSecs,
    /// Another label with a unit.
    #[metrics(unit = Unit::Bytes)]
    buffer_capacity: u64,
}

let labels = InfoLabels {
    version: "0.1.0",
    request_timeout: Duration::from_millis(100).into(),
    buffer_capacity: 1_024,
};
// will be exported as the following labels:
// { version="0.1.0", request_timeout_seconds="0.1", buffer_capacity="1024" }