vise

Derive Macro EncodeLabelValue

#[derive(EncodeLabelValue)]
{
    // Attributes available to this derive:
    #[metrics]
}

Derives the EncodeLabelValue trait for a type, which encodes a metric label value.

The type for which the trait is derived must not have generics (lifetimes, type params or const params). The macro can be configured using #[metrics()] attributes.

Container attributes

format

Type: string

Default value: {}

Specifies the format for the value as used in the format! macro etc. when encoding it to a label value. For example, {} means using Display.

rename_all

Type: string; one of lowercase, UPPERCASE, camelCase, snake_case, SCREAMING_SNAKE_CASE, kebab-case, SCREAMING-KEBAB-CASE

Specifies how enum variant names should be transformed into label values. This attribute can only be placed on enums in which all variants don’t have fields (aka C-style enums). Mutually exclusive with the format attribute.

Caveats:

rename_all assumes that original variant names are in PascalCase (i.e., follow Rust naming conventions).

rename_all requires original variant names to consist of ASCII chars.

Each letter of capitalized acronyms (e.g., “HTTP” in HTTPServer) is treated as a separate word. E.g., rename_all = "snake_case" will rename HTTPServer to h_t_t_p_server. Note that it is recommended to not capitalize acronyms (i.e., use HttpServer).

No spacing is inserted before numbers or other non-letter chars. E.g., rename_all = "snake_case" will rename Status500 to status500, not to status_500.

Variant attributes

name

Type: string

Specifies the name override for a particular enum variant when used with the rename_all attribute described above.

Examples

Default format

Label value using the default Display formatting; note that Display itself is derived.

use derive_more::Display;
use vise::EncodeLabelValue;

#[derive(Debug, Display, EncodeLabelValue)]
struct Method(&'static str);

Custom format

Label value using Hex formatting with 0 padding and 0x prepended.

#[derive(Debug, LowerHex, EncodeLabelValue)]
#[metrics(format = "0x{:02x}")]
struct ResponseType(u8);

Using rename_all on C-style enum

#[derive(Debug, EncodeLabelValue)]
#[metrics(rename_all = "snake_case")]
enum Database {
    Postgres, // renamed to "postgres"
    MySql, // renamed to "my_sql"
    #[metrics(name = "rocksdb")] // explicitly overrides the name
    RocksDB,
}