vise

Derive Macro Metrics

#[derive(Metrics)]
{
    // Attributes available to this derive:
    #[metrics]
}

Derives the Metrics trait for a type.

This macro must be placed on a struct with named fields and with no generics (lifetimes, type params or const params). Each field will be registered as metric or a family of metrics. The macro can be configured using #[metrics()] attributes.

Container attributes

prefix

Type: string

Specifies a common prefix for all metrics defined in the type. If specified, the prefix will be prepended together with a _ separator to a field name to get the metric name. Note that the metric name may be additionally transformed depending on the unit and metric type.

Field attributes

buckets

Type: expression evaluating to a type implementing Into<Buckets>

Specifies buckets for a Histogram or a Family of Histograms. This attribute is mandatory for these metric types and will result in a compile-time error if used on counters / gauges.

unit

Type: expression evaluating to Unit

Specifies unit of measurement for a metric. Note that specifying a unit influences the metric naming.

labels

Type: expression evaluating to an array [&'static str; _]

Specifies label names for a LabeledFamily. This attribute is mandatory for LabeledFamily and will result in a compile-time error if used with other metric types. The number of label names must match the number of label values in the LabeledFamily (i.e., its type).

Examples

See crate-level docs and other crate docs for the examples of usage.