vise

Module traits

Traits used for metric definitions, such as GaugeValue and HistogramValue.

Structs

AtomicIsizeWrapper

Thin wrapper around AtomicIsize used as atomic store for isize.

AtomicU64Wrapper

Thin wrapper around AtomicU64 used as atomic store for u64.

AtomicUsizeWrapper

Thin wrapper around AtomicUsize used as atomic store for usize.

LabelRef

Wrapper around a reference to labels proxying necessary trait implementations.

StaticLabelSet

Set of metric labels with label names known during compilation. Used as output in MapLabels implementations.

Enums

EncodedGaugeValue

Encoded value of a gauge.

Traits

EncodeLabelSet

Encodes a label set.

GaugeValue

Value of a Gauge.

HistogramValue

Value of a Histogram.

MapLabels

Maps a set of labels from the storage format (i.e., how labels are stored in a Family) to the encoding format, which is used when exporting metrics.